app/requirements.txt
aio-pika==9.5.4
aiodns==3.2.0
aioping==0.4.0
aiormq==6.8.1
aiosmb==0.4.11
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
asyauth==0.0.21
async-timeout==5.0.1
asyncpg==0.30.0
asysocks==0.2.13
bcrypt==4.0.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.4
gitignore_parser==0.1.11
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.5.0
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
minikerberos==0.4.4
multidict==6.1.0
oscrypto==1.3.0
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
pyasn1==0.6.1
pycares==4.5.0
pycparser==2.22
pycron==3.1.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pymod==0.2
pysmb==1.2.10
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
taskiq==0.11.9
taskiq-aio-pika==0.4.1
taskiq-dependencies==1.5.4
taskiq-fastapi==0.3.2
taskiq-redis==1.0.2
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
unicrypto==0.0.10
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.0
winacl==0.1.9
yarl==1.18.3
zipp==3.21.0
